requirements.txt
asgiref==3.3.1
Django==3.1.6
pytz==2021.1
sqlparse==0.4.1


factory-boy==3.2.0
djangorestframework==3.12.3
djangorestframework-datatables==0.6.0